requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-binance==1.0.19
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6